Рефакторинг: Вынес большую часть функционала из ConsoleApp.cs Выделил основную часть функционала из ConsoleApp.cs в отдельные файлы ProcessWord.cs и PathImage.cs, которые размещены в новой директории AppExtensions.
